// packages/common/src/stream.d.ts
export interface IReadable extends IEventEmitter {
    readable: boolean;
    read(size?: number): Buffer;
    setEncoding(encoding: string): this;
    pause(): this;
    resume(): this;
    isPaused(): boolean;
    pipe<T extends IWritable>(destination: T, options?: {
        end?: boolean | undefined;
    }): T;
    unpipe(destination?: IWritable): this;
    unshift(chunk: any, encoding?: string): void;
    push(chunk: any, encoding?: string): boolean;
    wrap(oldStream: IReadable): this;
    destroy?(error?: any): any;
    [Symbol.asyncIterator](): AsyncIterableIterator<any>;
}
export interface IEnd {
    end(cb?: () => void): this;
    end(data: any, cb?: (err?: Error | null | undefined) => void): this;
    end(data: any, encoding?: string, cb?: () => void): this;
}
export interface IEndable extends IEventEmitter, IEnd {
}
export interface IWritable extends IEndable {
    writable: boolean;
    write(data: any, cb?: (err?: Error | null | undefined) => void): boolean;
    write(data: any, encoding?: string, cb?: (err?: Error | null | undefined) => void): boolean;
}
// ...
export type ReadableLike<T> = T | (T & IReadable);
export type WritableLike<T> = T | (T & IWritable);